core: add `AllowHostDirVolumePlugin: true` to SCC
Adds AllowHostDirVolumePlugin to the go representation of the rook-ceph
SCC for using the hostPath. Also making it consistent with the yaml
representation.

From Openshift docs

For backwards compatibility, the usage of allowHostDirVolumePlugin overrides
settings in the volumes field. For example, if allowHostDirVolumePlugin is
set to false but allowed in the volumes field, then the hostPath
value will be removed from volumes.

Signed-off-by: Santosh Pillai <sapillai@redhat.com>