helm: set correct ingress endpoint protocol
In the helm chart, the ingress for the dashbard was always referring
to the http-dashboard port on the dashboard service. If ssl is enabled
the https-dashboard port must be specified, or else the dashboard
will not be available through the ingress.

The default is also changed from http to https when the dashboard
is installed through the cluster helm chart.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>