rgw: swift design document: configure swift in a new section
Instead of mixing configuration for different protocols in the gateway
section, depending on the type option, introduce a new section protocols for
API specific configuration.

This allows configuring s3 and swift in one RGW and at the same time
simplifies validation of the CRD and allows for a clean separation of
concerns.

Signed-off-by: Sebastian Riese <sebastian.riese@cloudandheat.com>